Rename compile.ts back to build.ts
